requirements_dev.txt
###################################### IMPORTANT #########################################
# use pip freeze > requirements_dev.text to lock the project dependencies after first run #
##########################################################################################
django-breeze
psycopg2
psycopg2-binary
whitenoise
python-decouple